Keyboard shortcut improvements
• CTRL-Shift-F now initiates a new “server search”

In the past, about 100 times a day I would be looking at SQL code and
want to search for a different object name. So I’d have to: 1) hit F7 to
switch back to the object search interface; 2) use the mouse to click
into the “Name” field; 3) delete whatever was there (usually with
CTRL-A); 4) start typing the new search param; 5) hit Enter. Now 80% of
that work can be done via one keyboard shortcut.

IOW, CTRL-Shift-F takes you directly to the Object Name search in yellow
above and does “select all” within that textbox. This works whether
you’re looking at the SQL Code view or the Object Search view.

• CTRL-M now toggles the poor-man’s SQL Formatting on/off for the active
SQL code window

This used to be CTRL-Shift-F, but I thought that it made more sense for
the “server search.” Because it’s Searching on the server, but it’s
Finding stuff. And “forMat” has an “M” in it. So it totally makes sense,
right???

• Database selection has been changed

The “All” checkbox for selecting which database(s) to search has been
replaced with “check all” and “uncheck all” buttons

A lot of times I would want to search “all databases except Miser” or
something like that. That was a major pain in the ass because obviously
I’d have to uncheck the “all” box, then individually click on every
other DB except miser. Or I might start out with 3 DB’s checked, then
later I wanted to search 2 different ones – this meant I had to slowly
scroll through the list, carefully checking or unchecking things as I
went.

By default all DB’s will be initially be checked so the default behavior
is the same – EVERYTHING is searched out of the gate. But in the
scenarios where you need to limit things, it should be much more
convenient.